Move and rename Scala files.
Intended naming convention for main modules:

 * ilc.feature.<feature>.Syntax      AST nodes for <feature>
 * ilc.feature.<feature>.Pretty      printing for <feature>
 * ilc.feature.<feature>.Evaluation  evaluation for <feature>
 * ...

A <feature> is a language feature, such as first-class functions,
integer arithmetics, strings, polymorphic bags, etc. The files in
the feature/<feature> hierarchy should contain traits, ready to
be mixed together with other traits for the same <feature> and/or
traits for other <feature>s.

 * ilc.feature/<Feature>  Packaged feature for easy import

For every feature, this file should contain a trait that mixes
together all traits for this feature.

 * ilc.language/<Language>

A <language> is a set of features, such as the pure lambda
calculus or Atlas. The files in the language hierarchy should
contain traits that mostly mix together the corresponding traits
from the Feature hierarchy.

Generally, if a file defines a trait, it should also define a
companion object, and the companion object should extend the
trait. This should be set up in such a way that the companion
object is a fully-defined language that contains essentially just
the feature defined by the trait.

Intended naming convention for test files:

 * A unit test suite for module <package>.<name> is called
   <package>.<name>Suite

Old-commit-hash: ac620a2f057e13a1690f062c66c4febb6615ca0e