language.Bacchus.Tools: add derivation test with fine-grained differences
A fine-grained difference is not a replacement value. For maps, it
stores keys to be deleted, key-value pairs to be inserted, and
values to be updated. For sums, the difference between Left(x)
and Left(y) is stored as a tagged difference between x and y,
and so is the difference between Right(x) and Right(y).

The tool `assertCorrectTwice` tests the correctness of a derivative
both on the result of diff-term and on a fine-grained difference.

Old-commit-hash: 8cd289be3312a3c569b2b54f2c0c58fa1f8a03da