[typed-scala] add feature.base.Derivation
This is a trait to spell out the obligations of Derivation of other
features. A Derivation trait should implement

- deltaType
- updateTerm
- diffTerm

for all types defined with that feature. A default `derive` method
based on `diffTerm` will be supplied. Since `diffTerm` operates
most likely by recomputation, it is a good idea for language features
to supply the derivatives of their constructors.

Old-commit-hash: 4885dc4c204e227a71368064b26b12617b268f70