Implement semi-safe incremental example generation
This heuristic (described below) is disabled by default because it's not fully safe.

The heuristic is: If the classfile responsible for generating an example
has not changed, skip regeneration. This gives a significant speedup
when running specific tests, and the speedup will get bigger as we add
more tests.

This heuristic is disabled by default because dependencies of the
generated file might have changed. If you enable this and the heuristic
fail, you can remove the generated file to force regeneration.

Old-commit-hash: 7033ffdc8e8a26a0c80615aa001d0ac59c441df9