Workaround "hang"
I claimed the benchmark for bags was hanging. In fact, it was just 2
orders of magnitude slower than for object-language maps. It seems the
high cost might be due to Scala's bagUnion, which uses boxing, unlike
the one for maps (which is inlined).

Hence:
* allow subclasses to override the collection sizes (by turning vals
  into def/lazy vals).
* reduce collection sizes for bags.

Old-commit-hash: b4182934f4e52a05ac6f2b461dd1750ed61c78f9