Speedup bags a bit by switching union to HashMap.merged
Currently, the benchmark runtime seems quadratic in the collection size.
This commit doesn't change that, but optimizes a hotspot by using
HashMap.merged, which allows to specify a merging function for combining
values (hence should be used for combineWith, except it does not allow
dropping elements) and is implemented more efficiently than
`++` (according to the documentation). It is also much more efficient in
my benchmarks.

Subsequent profiling shows that the hotspot is now the implementation of
bagNegate, which is used to compute differences. However, that hotspot
is probably best addressed by implementing better derivatives -
something which must be done anyway.

Old-commit-hash: 764b1524b2341df9b0dc507702a09e578ef760dc