Switch Bags to HashMap
This should have little effect on performance, but allows to use
HashMap.merged, which improves performance a lot, as shown in next
commit.

Note that the type annotation on bagUnion is needed to make the return
type more specific, as crazy as that should sound (reason: ++ expects a
GenTraversableOnce, and that information makes the return type of its
argument, a call to flatMap, less specific, because of a known but
already debated quirk of Scala type inference).

Old-commit-hash: cf1ecb6084dda4c5984e0354359cd3ba6d79325d