change the syntax of abelian maps to use abelian groups
`foldGroup` is renamed to `foldByHom` and takes two groups as
arguments. The value group of the input map is there for security:
If the derivative of `foldByHom` is given a change in terms of
a group other than the one declared as the domain of the homomorphism,
then it should recompute.

Old-commit-hash: fd0b7073642297cdd94a6c46cba5da3cbee4bd5d