implement abelian group syntax as constructor & projections
Reason: Current stability framework can't tell the higher-order
arguments of a `fold` that their arguments are stable if the thing
being folded over is stable. `FoldAbelianGroup` is to be implemented
as a meta-function in terms of projections for now.

Old-commit-hash: eb936df73d032b4d576d35feca4d70e899ae5cf8