lower "isNilChange" from meta- to object-level in derive(FoldGroup)
"isNilChange" on meta level:

-  s.hasStableArgument(0)

"isNilChange" on object level, for groups:

+ ifEqualGroups((_G, dG)) {
+   // _G receives nil change
+ } {
+   // _G receives non-nil change
+ }

Reason: Benchmarks need to get hold of the group used in the
program, say the additive integer group, in order to pass the
derivative those changes as it can digest. It can't do it if
the group is burried inside the program. It can only do it if
the program takes the group as part of the input, in which case
the group can't be statically guaranteed to be stable any more.
A dynamic test is necessary to select the fast derivative with
confidence.

Old-commit-hash: c411733e1d6600f2abdcf011b278c73f11f5148a