provide random input & changes in HistogramBenchmark
Input and changes are maps of random size to bags of random size,
with the total number of elements fixed. Because of that, the vast
majority of changes are insertions. Probability of update or
deletion approaches zero as input size approaches infinity.

Should we make changes less random to have more updates and
deletions in bags?

Old-commit-hash: ae27ab9fff129d36410f89d9b3c0ea5296274e7c