Show we can update indexes incrementally
Implement a generic GroupBy combinator in our language. Given how we
designed our primitives, this should be automatically efficient.

Moreover, a specialized version of this code is already part of our
evaluation. Looking at the differences, I don't think this can be much
slower (or, if it is, it should be easy to fix).

Old-commit-hash: 5fa6738bc6063e678f93cfd71e71a84d2141aede