Hide Scalatest's Symbol.apply(Any) pimped method
In this code we get a test failure because we pimp an apply method on
Symbol, but another pimped method takes precedence. The offender comes
because we inherit from
[org.scalatest.Matchers](https://github.com/scalatest/scalatest/blob/8f541bd1b6ecab15e8bcd05e7d05d23be607811c/src/main/scala/org/scalatest/Matchers.scala#L1679).

The goal of the implicit is to support syntax like `book should
have ('title ("A Tale of Two Cities"))`; since we don't use the syntax,
we can avoid inheriting from Matchers and avoid the whole issue.

Old-commit-hash: 530e81f2f2be5fa3f79bdf51fef9607776d701a4