Try making letBetaReduceRule more precise
Like for NbE, we'd need to handle functions nested inside Lets. But here
it would be harder, since we're not using HOAS or deBrujin.

Also, even if activate this rule, we don't get any benefit in the
output. Once again, that's possibly because we normalize in a fixpoint,
and normalization handles this better. But we undo the conversion to let
and redo it here... Hm. I might have a different idea.

Old-commit-hash: 1a8e2dfc0201c7318cf00ebac6a08d26651bc337