Make output of pretty include initial indentation
That's more regular. To prevent breakage, default indentation to be
initially 0, and cleanup how the default is set (by overriding an
extension point) and how the examples handle it (by changing it
dynamically as needed, so that the output does not change).

Old-commit-hash: 9630d4ba03ed59983957f07b4cfaab6ea517ffbb