Switch from Naturals to Integers in many tests
Quite a few tests were hard to switch because they manipulate changes
directly, but an interface hiding those details is not trivial to come
by. The proper answer is probably, again, to make more object-language
stuff also available in the meta-language, but that takes work (and it
is somewhat in tension with the goal of having a simple implementation).

Old-commit-hash: 455c1914ca4048809644996328ecef5f9fb712a8