Usable sbteclipse config and output
Now the output of sbteclipse seems more usable, and it's enough to run
`eclipse`. I still hand-corrected the output to remove the various empty
folders, but they are only a problem if they're in the project and don't
exist. In the recommended workflow, the Eclipse project is not checked
in, so everybody runs `sbt eclipse`, which will generate these stupid
empty folders. They are still stupid, but they will not break the build.

Hence, in next commit I'll remove the generated files, finally.

Old-commit-hash: f3a3b9b274181ee9ae0573c7c68c3a3706bbdfd7