Type inference: refactor/fix substitution on TVar
The current substitution would duplicate the handling for type variables
and not recurse into other types, which is bad.

In the context of type inference, triggering this bug is
impossible (because variable are typed by collectConstraints(UAbs) with
either type variables or type annotations, which do not contain type
variables since in practice never use them in client programs. But this
seems fragile, so let's fix it anyway.