ANormalForm: Abstract over state difference
The whole difference between CSE and non-CSE can indeed be encapsulated
by changing the state which is propagated, as done here.

Most of this commit just updates code threading the state and accessing
it.

The nontrivial bit was that CSE does an extra lookup - we ensure this
always fails in the non-CSE case.

This avoids giving the state a conditional type (with a Map interface in
one case, and without this interface in the other).

This commits also avoids representing and updating both kinds of
state (as done before).