A-normalization: ensure all application results are bound
That is, the innermost let body should not be an application node.

Maybe this does not matter for A-normalization, but it does matter for
the variant we're investigating, namely static caching.

The resulting code might be a tiny bit hacky: it ensures that the return
value of aNormalize is not an App node, so it's either atomic or a
lambda, but it does not make use of this assumption. However, this is
safe because aNormalizeTerm already knows to leave atomic terms
alone (if copyPropagation is enabled).