Static caching: try to adapt callers (WIP)
This is trying to adapt callers of a function to use only the first
element of the returned tuple.

However, right now I need to duplicate lots of code from
ANormalFormStateful to override some parts. That's because I'm mixed in
together with it, but I want to keep the original functionality also
available.

So, stopping here and changing the structure to use a bit more
composition.