Switch ANormalForm from inheritance to composition
This means a lot of changes; in particular, since we are using
path-dependent types, this requires switching to sharing constraints.

Once you know what to look for, it's possible to encode them, but type
inference does not really want to infer the right types.

This is partially "my fault" though - see next commit. Not that one
could guess.