Sharing constraints with less type annotations
Look at all those type annotations going away in the commit:

```
-  val aNormalizer: let.ANormalFormStateful { val mySyntax: outer.type } = new let.ANormalFormStateful {
-    protected val mySyntax: outer.type = outer
+  val aNormalizer = new let.ANormalFormStateful {
+    val mySyntax: outer.type = outer
```

The old type is now inferred automatically, including the refinement for
`mySyntax`. It turns out this did not happen before just because
`mySyntax` was protected. This sort-of makes sense, but the error
messages you get are less than helpful - in this situation, Scalac
should arguably give you a warning.

This commit also `mySyntax` public at the original declaration site:
that's not needed, but it ensures that all overriders become public.