Cleanup code sharing between ANormalFormStateful and AddCaches
Now that they don't need to be mixed together, we don't need to be
careful with name clashes, or with clashes on the state of the fresh
variable generator.

This also means that aNormalizeName now calls the old addCaches, as I
wanted, but this seems to cause uninteded breakage. So probably that was
not a good idea, but now at least I can control this better.