Actually fix substitution in A-normalization
The problem was that when descending into function bodies, the current
substitutions were not propagated inwards. Now they are propagated both
inwards (which is much better) and outwards - which is fixed in next
commit. (Here it still seems to work because there is not enough
shadowing to trigger the problem, but I'm not sure if this is
guaranteed).

This fixes the compile errors in clients and bigClients; the transformed
programs now give the correct results.

This also reverts the previous fix, which acted in the wrong place:
terms are traversed inwards, and when a term is saved the bindings it
uses have already been fully processed, so no further substitution is
needed. This is nice because it avoids doing a second pass over the
term.