Add AddCaches2: New attempt at static caching
This is much better, but I do eta-expansion of primitives at the wrong
time, so AddCaches2.extendReturns can't special-case them as it should.

AddCaches is commented out.

AddCaches2 is a separate component, not a mixin at all: it does not even
inherit from ANormalFormStateful.

Warning: Worksheet output generated with TypeChecking = false.