Speedup pathfinding (for 3rd submission)
This was (hopefully) in the third submission.

Problem: Path finding took too long after getting to the second corner,
and we faulted CycleLimit.
Analysis: it was trying to get to the opposite corner!
Solution: cycle corners in a different order, so things get faster.

And... it worked! (On the example map). Though it faults after we get to
the last corner, and doesn't ever progress.

Old-commit-hash: 22cfa8c3a03e6d19e36e2f3bcbb7ca460309457a