LM: fix search, rename filter to filterNot
filter was named according to the Ruby convention, so it negates the
predicate and does a Haskell/Scala filter. Rename it to filterNot.

I and Eric missed this when we copy-pasted and altered filter to
implement search, that's why it didn't work (I didn't test it though).
Fix search to not negate its predicate.

Old-commit-hash: b41408a2fe9202169c8bc5e50dc32e8271beb34b