Comment out aborted code
This code tries to represent CBPV terms as a subclass of normal terms.
That's a hack without a use case yet, but it's not clear to me how to
proceed after processing CBPV --- maybe we'll need such a hack, maybe
we'll just erase CBPV into lambda-calculus (and/or A-normal form).