Reveal bug in type inference through testing
This could be fixed in different ways (free variables should be recorded
in a context, either beforehand or during the scan).

But instead of reinventing the wheel, I should just switch the algorithm
from one for principal types (taking a context in input) to one for
principal typings (computing a context as output).