Make sure to output all needed caches WIP XXX
Currently, they can be overriden in cacheMap, leading to errors in the
generated output.

Alternatively, we could do CSE and reuse caches. But that seems an
orthogonal concern, and a tricky one since we have impure expressions.

A problem is whether we're going to reference the right cache, when we
output multiple caches for the same term.