Refactor Pretty to avoid cyclic dependencies
This refactoring ensures that `Pretty` doesn't depend on `Syntax`, but
the other way around. In particular, this prevents `Types` from
depending through `Pretty` on `Syntax`. I could alternatively use
self-types to express cyclic dependencies, but I'd need to alter lots of
traits to add the self-type annotation (since they are not inherited).