PolymorphicConstant.{apply ↦ tapply}
Related to #16.

We have too many methods fighting to be `apply`, take one out.
(Not overly sure tapply should be the final name, but this could be fixed
by `sed`).

Fixing the compile errors with Eclipse took only half an hour, mostly
waiting for Eclipse to recompile and show further errors.